requirements.txt
albumentations==2.0.8
comet_ml==3.55.0
coremltools==9.0
Flask==3.1.2
GitPython==3.1.40
GitPython==3.1.46
ipython==8.12.3
matplotlib==3.8.0
mss==10.1.0
nncf==2.19.0
numpy==2.4.2
onnx==1.20.1
onnxruntime==1.24.1
onnxslim==0.1.84
opencv_python==4.9.0.80
openvino==2025.4.1
paddle==1.2.4
pafy==0.5.5
pandas==3.0.0
Pillow==12.1.0
psutil==5.9.6
pycocotools==2.0.11
PyYAML==6.0.1
PyYAML==6.0.3
Requests==2.32.5
scipy==1.17.0
seaborn==0.11.1
setuptools==75.7.0
tensorflow==2.14.0
tensorflow_intel==2.14.0
tensorflowjs==4.22.0
tensorrt==10.15.1.29
tflite_runtime==2.14.0
tflite_support==0.4.4
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
tritonclient==2.65.0
ultralytics==8.3.58
ultralytics_thop==2.0.13
x2paddle==1.6.0
